+ Tweak Allen's recent changes so the code works on Darwin.  On Darwin   a uint64_t /= unsigned long long, so explicitly use uint64_t in the   Compare function signatures in place of 'unsigned long long.'

+++ src/c4/Compare.cc
@@ -86,7 +86,8 @@ bool check_global_equiv(int local_value)
  * \param local_value integer value to check against
  * \return true if equivalent across all processors; false if not
  */
-bool check_global_equiv(unsigned long long local_value)
+//bool check_global_equiv(unsigned long long local_value)
+bool check_global_equiv(uint64_t local_value)
 {
     
     const int node  = rtt_c4::node();
